+++ src/jobs/shared/util.shared.ts
