+++ tsconfig.json
@@ -32,8 +32,8 @@
     "paths": {
       "@interfaces/*": ["./src/interfaces/*"],
       "@middleware/*": ["./src/middleware/*"],
-      "@services/*": ["./src/services/*"],
-      "@service-utils/*": ["./src/services/shared/*"],
+      "@jobs/*": ["./src/jobs/*"],
+      "@jobs-utils/*": ["./src/jobs/shared/*"],
       "@database/*": ["./src/database/*"],
       "@models/*": ["./src/models/*"],
       "@routes/*": ["./src/routes/*"],
